sbsigntool: fix cross compilation flags contamination
Tools used during the build by the host have to be built using the host
compilation flags rather than the targets compilation flag. Otherwise,
the tools may not be able to run on the host during the build leading to
error such as:

| Building ccan_depends
| /bin/sh: line 1: tools/configurator/configurator: No such file or
directory

stemming from the fact that the binary built is set to use the target
interpretor (lets say /lib/ld-linux-x86-64.so.2) rather than the
interpretor available on the host.

Fix this by building the ccan tools using sbsigntool-native and reusing
them during the target build.