grub: add .sbat data section
The Secure Boot Advanced Targeting (SBAT) is a mechanism to allow the
revocation of components in the boot path by using generation numbers
embedded into the EFI binaries. The SBAT metadata is located in an .sbat
data section that has set of UTF-8 strings as comma-separated values
(CSV).

Add .sbat data section via --sbat option of grub-mkimage as shim
requires it since 15.3.

References:
https://github.com/rhboot/shim/blob/main/SBAT.md
https://www.gnu.org/software/grub/manual/grub/html_node/Secure-Boot-Advanced-Targeting.html

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>