seloader: support passing a buffer to be verified
Previously, GRUB passes a file path/name to SELoader for verification
request. That is working if kernel/initrd images are in the same
EFI partition as GRUB.

But for the scenario kernel/initrd images are in a different partition,
since seloader can not access file accross file system, GRUB will always
get a verfication failure when booting kernel and initrd, although they
are correctly signed.

So we add the capability to GRUB/SELoader to verify a buffer, rather
than passing a file path only.

Signed-off-by: Yunguo Wei <yunguo.wei@windriver.com>
Signed-off-by: Yi Zhao <yi.zhao@windriver.com>