grub-efi: refresh patches for grub-2.12
The EFI shim lock verifier is added to grub[1], so drop the patches of
our own shim verification implementation.

The shim lock verifier is enabled by default when GRUB_SIGN_VERIFY="1".
When UEFI_SELOADER="1", the MOK2 verification is enabled and the shim
lock verifier is disabled.

Refer to shim lock verifier implementation[2], add a patch to fallback
to legacy mode when MOK2 verification is enabled.

[1] http://git.savannah.gnu.org/cgit/grub.git/commit/?id=878398c1a39079a54c2836e27e9d1b0eb67a66cd
[2] http://git.savannah.gnu.org/cgit/grub.git/commit/?id=6425c12cd77ad51ad24be84c092aefacf0875089

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>