tpm2-totp: add recipe
This is a reimplementation of tpmtotp software for TPM 2.0 using the
tpm2-tss software stack. Its purpose is to attest the trustworthiness of
a device against a human using time-based one-time passwords (TOTP),
facilitating the Trusted Platform Module (TPM) to bind the TOTP secret
to the known trustworthy system state.

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>