tpm2-tss: Security fix for CVE-2024-29040
CVE-2024-29040:
The JSON Quote Info returned by Fapi_Quote has to be deserialized by
Fapi_VerifyQuote to the TPM Structure `TPMS_ATTEST`. For the field
`TPM2_GENERATED magic` of this structure any number can be used in the
JSON structure. The verifier can receive a state which does not
represent the actual, possibly malicious state of the device under test.
The malicious device might get access to data it shouldn't, or can use
services it shouldn't be able to.

Reference:
https://nvd.nist.gov/vuln/detail/CVE-2024-29040

Patch from:
https://github.com/tpm2-software/tpm2-tss/commit/c24d2d685814f9cc983af5fe2a731b04c7457c35

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>