efitools: improve the native build
The native build of efitools is only needed for the tools, add a patch
so that we can just build and install the tools.

This means we don't need to depend on gnu-efi-native anymore, and also
the native recipe doesn't end up shipping a build-host-targetting
LockDown.efi in the sysroot which is obviously incorrect.

Signed-off-by: Ross Burton <ross.burton@arm.com>