grub: move install of grub-efi.cfg to grub-bootconf
By default grub-efi has set RDEPENDS:${PN} = "virtual-grub-bootconf",
which is grub-bootconf package. It is used for generation of boot
configuration file.

Currently it is not being used for generation, but signing it, when
secure boot is enabled, should be run in this recipe instead grub-efi,
as grub.cfg is part of it.

Signed-off-by: Piotr Łobacz <p.lobacz@welotec.com>