cryptfs-tpm2: disable linker GCS warning on aarch64
Fixes:
libcryptfs-tpm2.so: warning: GCS is required by -z gcs, but this shared
library lacks the necessary property note. The dynamic loader might not
enable GCS or refuse to load the program unless all the shared library
dependencies have the GCS marking.

Gcc15 has added _CHKFEAT_GCS in aarch64[1]. It's a distro-wide setting
to warn when one of inputs does not have GCS. This would not be a
problem but libcryptfs-tpm2 links with -fatal-warnings, so the build
explodes.

[1] https://gcc.gnu.org/git/?p=gcc.git;a=commit;h=5a6af707f0af

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>